+++ extension/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "al-translation-cleaner" extension will be documented
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [1.0.0] - 2026-03-10
+
+### Added
+
+* Command:Write Translations from All Xliff Files to AL File Comments (See [README.md](README.md#atc-write-translations-from-all-xliff-files-to-al-file-comments) for more info)
+
+### Fixed
+
+* Translations are not written into Locked Lines
+
 ## [1.0.0-beta-1] - 2026-03-10
 
 ### Added
@@ -24,12 +34,12 @@ All notable changes to the "al-translation-cleaner" extension will be documented
 
 ### Added
 
-- Command: Find Translations in Xliff File (See [README.md](README.md#atc-find-translations-in-xliff-file) for more info)
+- Command: Find Translations in Xliff File (See [README.md](README.md#atc-write-translations-in-current-xliff-file-to-al-file-comments) for more info)
 
 ## [0.0.1-beta] - 2026-03-09
 
 ### Added
 
-- Command: Find Translations in AL File (See [README.md](README.md#atc-find-translations-in-file) for more info)
+- Command: Find Translations in AL File (See [README.md](README.md#atc-write-translations-to-comments-in-current-al-file) for more info)
 - Setting: ATC.TranslationMethod
 - Setting: ATC.languageMapping